@@ -1,17 +1,15 @@
 name: simple_rich_text
 description: Display a Dart string in Flutter using simple rich text characters.
-version: 1.0.35        #NOEDIT
+version: 2.0.36        #NOEDIT
 homepage: https://github.com/remoteportal/simple_rich_text
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.1.8
-  simple_navigation:
-    git: https://github.com/remoteportal/simple_navigation.git
+  simple_navigation: ^1.0.31
 
 dev_dependencies:
   flutter_test:
